Move size to berlin container
